refactor: removed logging from helper class and implemented it in the functions and resolved comments
