+++ src/BIAnalyticsManagementService/GetParticipantScreeningEpisode/GetParticipantScreeningEpisode.cs
@@ -1,4 +1,3 @@
-using System.Globalization;
 using System.Net;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
@@ -21,17 +20,19 @@ public class GetParticipantScreeningEpisode
     public async Task<HttpResponseData> Run([HttpTrigger(AuthorizationLevel.Anonymous, "get")] HttpRequestData req)
     {
         _logger.LogInformation("GetParticipantScreeningEpisode start");
-        var paginationHelper = new PaginationHelper(_logger);
+        var paginationHelper = new PaginationHelper();
 
-        if (!paginationHelper.TryValidatePaginationQuery(req, out int page, out int pageSize, out DateTime startDate, out DateTime endDate))
+        if (!paginationHelper.TryValidatePaginationQuery(req.Query, out int page, out int pageSize, out DateTime startDate, out DateTime endDate, out string errorMessage))
         {
             var errorResponse = req.CreateResponse(HttpStatusCode.BadRequest);
+            errorResponse.WriteString(errorMessage);
             return errorResponse;
         }
 
-        var requestHandler = new PaginationHelper(_logger);
+        var requestHandler = new PaginationHelper();
         string baseUrl = Environment.GetEnvironmentVariable("GetParticipantScreeningEpisodeDataUrl");
         string url = requestHandler.BuildUrl(baseUrl, page, pageSize, startDate, endDate);
+        _logger.LogInformation("Requesting URL: {Url}", url);
 
         try
         {
